updates to get celery task working
