+++ requirements
@@ -24,3 +24,4 @@ boto3==1.34.89
 botocore==1.34.89
 celery>=5.2,<6.0
 redis>=4.0,<5.0
+django-celery-results>=2.5.1
